+++ Contents/mods/B42 PZLinux/42/media/lua/shared/TimedActions/ISTakeTheCargoAction.lua
@@ -7,12 +7,12 @@ function ISTakeTheCargoAction:isValid()
 end
 
 function ISTakeTheCargoAction:waitToStart()
-    self.character:faceThisObject(self.item)
+    if self.item then self.character:faceThisObject(self.item) end
 	return self.character:shouldBeTurning()
 end
 
 function ISTakeTheCargoAction:update()
-    self.character:faceThisObject(self.item)
+    if self.item then self.character:faceThisObject(self.item) end
 end
 
 function ISTakeTheCargoAction:start()
@@ -26,9 +26,7 @@ function ISTakeTheCargoAction:stop()
 end
 
 function ISTakeTheCargoAction:perform()
-    PZLinuxRequestContractWorldEvent(self.character, "takeCargo", {
-        target = PZLinuxGetWorldObjectReference(self.item),
-    }, function(result)
+    PZLinuxRequestContractWorldEvent(self.character, "takeCargo", {}, function(result)
         if not result or not result.ok then return end
         local helicopterHandler = rawget(_G, "testHelicopter")
         if type(helicopterHandler) == "function" then
@@ -40,12 +38,15 @@ function ISTakeTheCargoAction:perform()
     ISBaseTimedAction.perform(self)
 end
 
-function ISTakeTheCargoAction:new(character, item)
+function ISTakeTheCargoAction:new(character, item, x, y, z)
     local o = ISBaseTimedAction.new(self, character)
     setmetatable(o, self)
     self.__index = self
     o.character = character
     o.item = item
+    o.x = x
+    o.y = y
+    o.z = z
     o.stopOnWalk = true
     o.maxTime = 1000
     return o